% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covmat_pcp.R
\name{covmat_pcp}
\alias{covmat_pcp}
\title{Parallel coordinate plot of a covariance matrix distribution}
\usage{
covmat(x, xnames=NULL, colvar = 'dep')
}
\arguments{
\item{x}{a list of covariance matrices}

\item{xnames}{a (optional) character vector with names for x}

\item{colorvar}{color variable of each covariance. Available options are effective dependence, 'dep', effective variance 'var', or an user privided index}
}
\description{
Parallel coordinate plot of a covariance matrix distribution
}
\examples{
smpcovs <- rIW(n=100, d = 15, R = diag(4) )
covmat_pcp(x = smpcovs)

}
