% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PriorCovmatrix-package.R
\docType{package}
\name{PriorCovmatrix-package}
\alias{PriorCovmatrix-package}
\alias{PriorCovmatrix}
\title{The 'PriorCovmatrix' package.}
\description{
Simulate and visualize covariance matrix distributions
}
\references{
Stan Development Team (2018). RStan: the R interface to Stan. R package version 2.17.3. http://mc-stan.org
}
