% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rSS.R
\name{rSS}
\alias{rSS}
\title{Random generation of covariance matrix using Separation Strategy}
\usage{
rSS(n, k prior_cor = 'lkj', prior_sg ='ln', eta = k+1, R = diag(k), sigma_mu=0, sigma_sc=1)
}
\arguments{
\item{n}{number of observations}

\item{k}{covariance matrix dimension}

\item{prior_cor}{prior for correlation matrix}

\item{prior_sg}{prior for marginal variances}

\item{eta}{degrees of freedom IW parameter}

\item{R}{matrix IW parameter}

\item{sigma_mu}{location of variances}

\item{sigma_sc}{scale of variances}
}
\description{
Random generation of covariance matrix using Separation Strategy
}
\examples{
rSS(n=2, k = 5)

}
