% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simula_IW.R
\name{simula_IW}
\alias{simula_IW}
\title{Generates random draws from inverse-Wishart distribution}
\usage{
simula_IW(n, d, R)
}
\arguments{
\item{n}{number of observations}

\item{nu}{degrees of freedom}

\item{R}{positive definite matrix}
}
\description{
Generates random draws from inverse-Wishart distribution
}

