% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simula_SIW.R
\name{simula_SIW}
\alias{simula_SIW}
\title{Generates random draws from scaled inverse-Wishart distribution}
\usage{
simula_SIW(n, d, R)
}
\arguments{
\item{n}{number of observations}

\item{d}{matrix dimension}

\item{nu}{degrees of freedom}

\item{xi}{common standard deviation of marginal variances}
}
\description{
Generates random draws from scaled inverse-Wishart distribution
}

