\name{PriorCovmatrix-package}
\alias{PriorCovmatrix-package}
\alias{PriorCovmatrix}
\docType{package}
\title{
\packageTitle{PriorCovmatrix}
}
\description{
\packageDescription{PriorCovmatrix}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{PriorCovmatrix}
\packageIndices{PriorCovmatrix}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{PriorCovmatrix}

Maintainer: \packageMaintainer{PriorCovmatrix}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
